Convert site to Astro
