--- .eslintrc.json
@@ -1,12 +1,0 @@
-{
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "./tsconfig.json"
-  },
-  "plugins": ["@typescript-eslint"],
-  "extends": ["next/core-web-vitals", "plugin:@typescript-eslint/recommended"],
-  "rules": {
-    "@next/next/no-img-element": "off",
-    "@typescript-eslint/no-non-null-assertion": "off"
-  }
-}
