--- feedbacks.md
@@ -1,29 +1,0 @@
-## Feedbacks to work on 
-
-1. beginning animation with the logos are a bit distracting, consider using a horizontal fly-in with opacity fade in
-1. title font and content font sizes should be consistent
-1. photo on the left instead of the right (imo)
-1. add more clear separation between pages and contents
-1. center the content you want users to see, in this case your about me
-1. smaller navbar, and also center your elements as you don't have a lot for now
-1. scaling reaction is a bit overdone, tone it down a bit... maybe
-1. use dropshadows for better contrast and as an indication to the fact that this is an interactable element
-1. a bit more border radius on the boxes (imo)
-1. watch your text box width, you should be able to fit at least 3-5 words per line. Decrease your thumbnail size if it's necessary.
-1. carve out a dedicated area for content stats (like counts, dates, etc...)
-1. add your copyright at the bottom
-1. a contact me form would do wonders
-1. some sort of page navigation, but it's not strictly necessary. I don't have one, but it might be nice for you to have one
-1. actually explain in the "What do I do"... what do you do?
-1. avoid using gradients on your navbar, it's really hard to get right for bigger elements. Instead, use a high opacity (but not opaque) background color, and add a backdrop filter with blur(13px) or so on it. It will look fantastic.
-1. stretch your contents our a bit more to fill the void on the sides
-1. give the page a proper conclusion rather than just ending it there. This goes back to the contact form and stuff.
-1. use more font varieties. I know it's not Poppins, but try giving your title and body different fonts. It makes them stand out more.
-1. good mobile layout, just watch the centering and stuff a bit more
-1. mobile link for blog on the nav just disappears?
-1. watch horizontal overflow on your mobile view. This will lead to an awful UI experience
-1. disable hover on mobile. It will cause problems when people actually use it on their phone (unless it's intentional)
-1. good mobile card layouts
-1. get a dedicated social section for mobile, so that it's easier to view and click on 
-1. better twitter cards
-1. Link hover without `inline-block` [reference](https://github.com/genicsblog/theme-files/pull/26)
