+++ package.json
@@ -1,46 +1,28 @@
 {
-  "name": "usmans.me",
-  "version": "0.1.0",
+  "name": "@example/minimal",
+  "type": "module",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "astro dev",
+    "start": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview",
+    "astro": "astro"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.2",
-    "@fontsource/montserrat": "^4.5.13",
-    "@fontsource/sen": "^4.5.9",
-    "@lottiefiles/react-lottie-player": "^3.5.0",
-    "axios": "^1.2.1",
-    "framer-motion": "^7.10.2",
-    "graphql": "^16.6.0",
+    "@astrojs/react": "^2.0.2",
+    "@astrojs/tailwind": "^3.0.1",
+    "@lottiefiles/react-lottie-player": "^3.5.2",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "astro": "^2.0.14",
+    "astro-seo": "^0.7.1",
+    "framer-motion": "^9.0.4",
     "millify": "^5.0.1",
-    "next": "12.3.1",
-    "next-seo": "^5.15.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
-    "swr": "^2.0.0",
-    "zod": "^3.18.0"
-  },
-  "devDependencies": {
-    "@types/node": "18.0.0",
-    "@types/react": "18.0.14",
-    "@types/react-dom": "18.0.5",
-    "@typescript-eslint/eslint-plugin": "^5.33.0",
-    "@typescript-eslint/parser": "^5.33.0",
-    "autoprefixer": "^10.4.13",
-    "eslint": "8.22.0",
-    "eslint-config-next": "12.3.1",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.4",
-    "prettier-plugin-tailwindcss": "^0.2.3",
-    "tailwindcss": "^3.2.7",
-    "typescript": "^4.8.4"
-  },
-  "ct3aMetadata": {
-    "initVersion": "6.5.2"
+    "tailwindcss": "^3.2.7"
   }
 }
