+++ src/components/Nav.astro
@@ -1,0 +1,21 @@
+<div class='w-full bg-primary-dark'>
+  <div class='p-5'>
+    <div class='flex items-center justify-center md:justify-between'>
+      <h1 class='text-center text-4xl font-semibold text-primary-light'>
+        Usman Sabuwala
+      </h1>
+      <div class='hidden gap-5 md:flex'>
+        <a
+          class='text-xl text-primary-light no-underline'
+          href='https://blog.usmans.me'
+        >
+          Blog
+        </a>
+
+        <a class='text-xl text-primary-light no-underline' href='/gists'>
+          Gists
+        </a>
+      </div>
+    </div>
+  </div>
+</div>
