+++ src/components/PostCard.tsx
@@ -1,0 +1,48 @@
+import millify from 'millify';
+import { FiThumbsUp } from 'react-icons/fi';
+import { MdComment } from 'react-icons/md';
+import type { Post } from '../types';
+
+export default function PostCards({ posts }: { posts: Post[] }) {
+  return posts.map(post => <PostCard key={post.cuid} post={post} />);
+}
+
+function PostCard({ post }: { post: Post }) {
+  async function sendBlogClickMessage() {
+    await fetch(`/api/sendDiscordMessage?name=Blog - ${post.title}`);
+  }
+
+  return (
+    <a
+      onClick={sendBlogClickMessage}
+      href={'https://blog.usmans.me/' + post.slug}
+      target='_blank'
+      rel='noreferrer'
+      className='h-full'
+    >
+      <div className='h-full max-w-sm cursor-pointer overflow-hidden rounded-lg bg-card-bg transition-colors hover:bg-opacity-50'>
+        <img
+          src={
+            'https://res.cloudinary.com/demo/image/fetch/f_auto/' +
+            post.coverImage
+          }
+          alt={post.title}
+          title={post.title}
+          className='w-full'
+        />
+        <div className='h-full p-6 '>
+          <h4 className='text-xl font-semibold text-white'>{post.title}</h4>
+          <p className='mt-2 text-lg '>
+            <p className='flex items-center gap-2 text-gray-300'>
+              <MdComment /> {post.replyCount + post.responseCount} Comments
+            </p>
+
+            <p className='flex items-center gap-2 text-gray-300'>
+              <FiThumbsUp /> {millify(post.totalReactions)} Likes
+            </p>
+          </p>
+        </div>
+      </div>
+    </a>
+  );
+}
