+++ src/components/Separator.astro
@@ -1,0 +1,3 @@
+<div class='my-7 grid place-items-center'>
+  <div class='h-1 w-10/12 bg-separator'></div>
+</div>
