--- src/components/Separator.tsx
@@ -1,9 +1,0 @@
-function Separator() {
-  return (
-    <div className='my-7 grid place-items-center'>
-      <div className='h-1 w-10/12 bg-separator'></div>
-    </div>
-  );
-}
-
-export { Separator };
