+++ src/components/SocialLink.tsx
@@ -1,5 +1,17 @@
 import { motion, TargetAndTransition } from 'framer-motion';
-import { ReactNode } from 'react';
+import type { ReactNode } from 'react';
+import { socialLinks } from '../utils/socialLinks';
+
+function SocialLinks() {
+  return socialLinks.map(socialLink => (
+    <SocialLink
+      key={socialLink.link}
+      link={socialLink.link}
+      icon={socialLink.icon}
+      name={socialLink.name}
+    />
+  ));
+}
 
 const SocialLink = (props: { link: string; icon: ReactNode; name: string }) => {
   const sendMessage = async () => {
@@ -24,4 +36,4 @@ const SocialLink = (props: { link: string; icon: ReactNode; name: string }) => {
   );
 };
 
-export { SocialLink };
+export { SocialLinks };
