+++ src/components/page-components/Hero.astro
@@ -1,0 +1,29 @@
+---
+import SectionHeading from '../SectionHeading.astro';
+import { SocialLinks } from '../SocialLink';
+---
+
+<div class='mt-8'>
+  <div
+    class='flex flex-col-reverse items-center justify-center gap-3 md:flex-row'
+  >
+    <div class='flex flex-col gap-5'>
+      <SectionHeading>Hi, I am Usman!</SectionHeading>
+      <p class='text-lg text-white'>
+        Hey all! My name is Usman! I am 17 years old and learning web
+        development and programming. I also have a YouTube channel on the
+        subject of programming and mostly web development.
+      </p>
+
+      <p class='text-lg text-white'>
+        Open to remote work opportunities. Feel free to contact me.
+      </p>
+
+      <div class='flex gap-5 items-center'>
+        <SectionHeading>Contact me:</SectionHeading>
+        <SocialLinks client:load />
+      </div>
+    </div>
+    <img src='/profilePic.jpg' alt='Usman' class='max-w-xs rounded-full' />
+  </div>
+</div>
