--- src/pages/_document.tsx
@@ -1,19 +1,0 @@
-import Document, { Head, Html, Main, NextScript } from 'next/document';
-
-import { AppConfig } from '@/utils/AppConfig';
-
-class MyDocument extends Document {
-  render() {
-    return (
-      <Html lang={AppConfig.locale}>
-        <Head />
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
-
-export default MyDocument;
