--- src/styles/global.css
@@ -1,11 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-html {
-  overflow-x: hidden;
-}
-
-*::selection {
-  background-color:  rgba(242, 148, 146, 0.3);
-}
