--- src/utils/AppConfig.ts
@@ -1,6 +1,0 @@
-export const AppConfig = {
-  site_name: 'Usman',
-  title: 'Usman Sabuwala - Web Developer in India.',
-  description: '🌐 my online home',
-  locale: 'en',
-};
