--- src/utils/apolloClient.ts
@@ -1,8 +1,0 @@
-import { ApolloClient, InMemoryCache } from '@apollo/client';
-
-export const client = (uri: string) => {
-  return new ApolloClient({
-    uri,
-    cache: new InMemoryCache(),
-  });
-};
