+++ src/utils/fetchPosts.ts
@@ -1,24 +1,4 @@
-import { gql } from '@apollo/client';
-import axios from 'axios';
-
-import { client } from './apolloClient';
-
-export interface Post {
-  totalReactions: number;
-  coverImage: string;
-  cuid: string;
-  dateAdded: string;
-  slug: string;
-  title: string;
-  responseCount: number;
-  replyCount: number;
-}
-
-export interface LikesAndComments {
-  totalReactions: number;
-  responseCount: number;
-  replyCount: number;
-}
+import type { Post } from '../types';
 
 interface Data {
   user: {
@@ -28,41 +8,12 @@ interface Data {
   };
 }
 
-async function getDevToPosts() {
-  const { data } = await axios.get(
-    'https://dev.to/api/articles?username=maxprogramming&per_page=5'
-  );
-  return (data as any[]).map((d: any) => ({
-    url: d.url,
-    canonicalUrl: d.canonical_url,
-  }));
-}
-
-export async function getPostLikesAndComments() {
-  const hashnodeClient = client('https://api.hashnode.com');
-  const { data } = await hashnodeClient.query<any>({
-    query: gql`
-      query GetPosts {
-        user(username: "usmanwrites") {
-          publication {
-            posts {
-              totalReactions
-              responseCount
-              replyCount
-            }
-          }
-        }
-      }
-    `,
-  });
-  const likesAndComments: LikesAndComments[] = data.user.publication.posts;
-  return likesAndComments;
-}
-
-async function getHashnodePosts() {
-  const hashnodeClient = client('https://api.hashnode.com');
-  const { data } = await hashnodeClient.query<Data>({
-    query: gql`
+export default async function fetchPosts() {
+  const response = await fetch('https://api.hashnode.com', {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({
+      query: `
       query GetPosts {
         user(username: "usmanwrites") {
           publication {
@@ -79,18 +30,11 @@ async function getHashnodePosts() {
           }
         }
       }
-    `,
+      `,
+    }),
   });
-  return data;
-}
 
-export async function getPosts() {
-  const [hashnodePosts, devToPosts] = await Promise.all([
-    getHashnodePosts(),
-    getDevToPosts(),
-  ]);
-
-  const { posts } = hashnodePosts.user.publication;
-
+  const { data }: { data: Data } = await response.json();
+  const { posts } = data.user.publication;
   return posts;
 }
